Major overhaul to the background system
The new background system is based on agdq16-layout's background system, is completely modular, and overall is better for the stream and making things look nice.

In the future, the backgrounds could be swapped to an animated version using -webkit-mask-image, but at the moment I'm not sure entirely how much that could affect performance on my normal streaming hardware. Currently it takes a toll on my poor laptop, so I can't test much about it yet, but when I get home it'll be worked on.